+++ scripts/bootstrap/debian/2_packages.sh
@@ -4,16 +4,24 @@ echo '##############################################'
 echo "Starting ${0}.."
 set -x
 
+if [[ -n "$MANAGE_SOURCES_LIST" ]]; then #TODO make it more configurable
+cat <<EOF > /etc/apt/sources.list
+deb http://ftp.debian.org/debian ${OS_CODENAME} main contrib non-free
+deb-src http://ftp.debian.org/debian ${OS_CODENAME} main contrib non-free
+
+deb http://security.debian.org/ ${OS_CODENAME}/updates main contrib non-free
+deb-src http://security.debian.org/ ${OS_CODENAME}/updates main contrib non-free
+
+deb http://ftp.debian.org/debian ${OS_CODENAME}-updates main contrib non-free
+deb-src http://ftp.debian.org/debian ${OS_CODENAME}-updates main contrib non-free
+
+deb http://ftp.debian.org/debian ${OS_CODENAME}-backports main contrib non-free
+deb-src http://ftp.debian.org/debian ${OS_CODENAME}-backports main contrib non-free
+EOF
+fi
+
 apt-get -q update
 DEBIAN_FRONTEND=noninteractive apt-get -yV upgrade
-DEBIAN_FRONTEND=noninteractive apt-get -yV install apt-transport-https linux-headers-$(uname -r) bind9-host
-
-# update package index on boot
-cat <<EOF > /etc/init/refresh-apt.conf
-description "update package index"
-start on networking
-task
-exec /usr/bin/apt-get -q update
-EOF
+DEBIAN_FRONTEND=noninteractive apt-get -yV install apt-transport-https bind9-host wget
 
 echo "Finishing ${0}.."